+++ src/main/__tests__/cordova.spec.ts
@@ -1,30 +1,41 @@
 import fetchMock from 'jest-fetch-mock'
 
 import { toQueryString } from '../../utils/queryString'
-import ApiClient from '../apiClient'
+import { initCordovaCallbackIfNecessary } from '../cordovaHelper'
+import { createHttpClient } from '../httpClient'
 import createEventManager from '../identityEventManager'
+import OauthClient from '../oAuthClient'
 import createUrlParser from '../urlParser'
 import { headersTest } from './helpers/identityHelpers'
 import { defaultScope, mockPkceValues } from './helpers/oauthHelpers'
 import { defineWindowProperty, mockWindowCrypto } from './helpers/windowHelpers'
 
 const clientId = 'kqIJE'
-const domain = 'local.reach5.net'
+const baseUrl = 'https://local.reach5.net'
+const baseIdentityUrl = `${baseUrl}/identity/v1`
+const http = createHttpClient({
+  baseUrl: baseIdentityUrl,
+  language: 'en',
+  acceptCookies: false,
+})
 
 function apiClientAndEventManager() {
   const eventManager = createEventManager()
-  const client = new ApiClient({
+  const client = new OauthClient({
     config: {
       clientId,
-      domain,
+      baseUrl,
       language: 'en',
       sso: false,
       pkceEnforced: false,
       isPublic: true,
     },
+    http,
     eventManager,
-    urlParser: createUrlParser(eventManager),
   })
+  const urlParser = createUrlParser(eventManager)
+  initCordovaCallbackIfNecessary(urlParser)
+
   return { client, eventManager }
 }
 
@@ -80,7 +91,7 @@ describe('signup', () => {
     })
 
     // Then
-    expect(signupCall).toHaveBeenCalledWith(`https://${domain}/identity/v1/signup-token`, {
+    expect(signupCall).toHaveBeenCalledWith(`${baseIdentityUrl}/signup-token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -150,7 +161,7 @@ describe('signup', () => {
     })
 
     // Then
-    expect(signupCall).toHaveBeenCalledWith(`https://${domain}/identity/v1/signup-token`, {
+    expect(signupCall).toHaveBeenCalledWith(`${baseIdentityUrl}/signup-token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -272,7 +283,7 @@ describe('loginWithPassword', () => {
     await client.loginWithPassword({ email, password })
 
     // Then
-    expect(passwordLoginCall).toHaveBeenCalledWith(`https://${domain}/oauth/token`, {
+    expect(passwordLoginCall).toHaveBeenCalledWith(`${baseUrl}/oauth/token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -325,7 +336,7 @@ describe('loginWithPassword', () => {
     await client.loginWithPassword({ phoneNumber, password })
 
     // Then
-    expect(passwordLoginCall).toHaveBeenCalledWith(`https://${domain}/oauth/token`, {
+    expect(passwordLoginCall).toHaveBeenCalledWith(`${baseUrl}/oauth/token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -384,7 +395,7 @@ describe('loginWithPassword', () => {
     })
 
     // Then
-    expect(passwordLoginCall).toHaveBeenCalledWith(`https://${domain}/oauth/token`, {
+    expect(passwordLoginCall).toHaveBeenCalledWith(`${baseUrl}/oauth/token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -445,7 +456,7 @@ describe('loginWithPassword', () => {
 
     // Then
 
-    expect(passwordLoginCall).toHaveBeenCalledWith(`https://${domain}/oauth/token`, {
+    expect(passwordLoginCall).toHaveBeenCalledWith(`${baseUrl}/oauth/token`, {
       method: 'POST',
       headers: headersTest.jsonAndDefaultLang,
       body: JSON.stringify({
@@ -530,7 +541,7 @@ describe('loginWithSocialProvider', () => {
     expect(window.cordova.plugins?.browsertab?.openUrl).toHaveBeenCalledTimes(1)
 
     expect(calledUrl).toEqual(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -580,7 +591,7 @@ describe('loginWithSocialProvider', () => {
     expect(window.cordova.plugins?.browsertab?.openUrl).not.toHaveBeenCalled()
 
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -620,7 +631,7 @@ describe('loginWithSocialProvider', () => {
     expect(window.cordova.plugins?.browsertab?.openUrl).not.toHaveBeenCalled()
 
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -648,7 +659,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -676,7 +687,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -708,7 +719,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'code',
@@ -742,7 +753,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'code',
@@ -774,7 +785,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',
@@ -804,7 +815,7 @@ describe('loginWithSocialProvider', () => {
 
     // Then
     expect(window.cordova.InAppBrowser?.open).toHaveBeenCalledWith(
-      `https://${domain}/oauth/authorize?` +
+      `${baseUrl}/oauth/authorize?` +
         toQueryString({
           client_id: clientId,
           response_type: 'token',