+++ src/main/main.ts
@@ -1,36 +1,50 @@
 import { toQueryString } from '../utils/queryString'
-import ApiClient, {
-  EmailVerificationParams,
-  LoginWithCredentialsParams,
-  LoginWithPasswordParams,
-  PasswordlessParams,
-  PhoneNumberVerificationParams,
-  RefreshTokenParams,
+import { AuthOptions } from './authOptions'
+import { AuthResult } from './authResult'
+import { initCordovaCallbackIfNecessary } from './cordovaHelper'
+import { createHttpClient, rawRequest } from './httpClient'
+import createEventManager, { Events } from './identityEventManager'
+import MfaClient, {
   RemoveMfaEmailParams,
   RemoveMfaPhoneNumberParams,
-  RequestPasswordResetParams,
-  SignupParams,
   StartMfaEmailRegistrationParams,
   StartMfaEmailRegistrationResponse,
   StartMfaPhoneNumberRegistrationParams,
+  StartMfaPhoneNumberRegistrationResponse,
   StepUpParams,
-  TokenRequestParameters,
-  UpdateEmailParams,
-  UpdatePasswordParams,
   VerifyMfaEmailRegistrationParams,
   VerifyMfaPasswordlessParams,
   VerifyMfaPhoneNumberRegistrationParams,
-  VerifyPasswordlessParams,
-} from './apiClient'
-import { AuthOptions } from './authOptions'
-import { AuthResult } from './authResult'
-import { rawRequest } from './httpClient'
-import createEventManager, { Events } from './identityEventManager'
+} from './mfaClient'
 import { MFA, OpenIdUser, PasswordlessResponse, Profile, RemoteSettings, SessionInfo } from './models'
+import OAuthClient, {
+  LoginWithCredentialsParams,
+  LoginWithCustomTokenParams,
+  LoginWithPasswordParams,
+  LogoutParams,
+  RefreshTokenParams,
+  SignupParams,
+  SingleFactorPasswordlessParams,
+  TokenRequestParameters,
+  VerifyPasswordlessParams,
+} from './oAuthClient'
+import ProfileClient, {
+  EmailVerificationParams,
+  GetUserParams,
+  PhoneNumberVerificationParams,
+  RequestPasswordResetParams,
+  UnlinkParams,
+  UpdateEmailParams,
+  UpdatePasswordParams,
+  UpdatePhoneNumberParams,
+  UpdateProfileParams,
+  VerifyPhoneNumberParams,
+} from './profileClient'
 import createUrlParser from './urlParser'
+import WebAuthnClient from './webAuthnClient'
+import CredentialsResponse = MFA.CredentialsResponse
 import { DeviceCredential, LoginWithWebAuthnParams, SignupWithWebAuthnParams } from './webAuthnService'
 import StepUpResponse = MFA.StepUpResponse
-import MfaCredentialsResponse = MFA.CredentialsResponse
 
 export { AuthOptions } from './authOptions'
 export { AuthResult } from './authResult'
@@ -43,26 +57,40 @@ export interface Config {
   language?: string
 }
 
+export type ApiClientConfig = {
+  clientId: string
+  language?: string
+  scope?: string
+  sso: boolean
+  pkceEnforced: boolean
+  isPublic: boolean
+  baseUrl: string
+}
+
 export type Client = {
   addNewWebAuthnDevice: (accessToken: string, friendlyName?: string) => Promise<void>
   checkSession: (options?: AuthOptions) => Promise<AuthResult>
   checkUrlFragment: (url: string) => boolean
   exchangeAuthorizationCodeWithPkce: (params: TokenRequestParameters) => Promise<AuthResult>
+  getMfaStepUpToken: (params: StepUpParams) => Promise<StepUpResponse>
   getSessionInfo: () => Promise<SessionInfo>
   getSignupData: (signupToken: string) => Promise<OpenIdUser>
-  getUser: (params: { accessToken: string; fields?: string }) => Promise<Profile>
+  getUser: (params: GetUserParams) => Promise<Profile>
+  listMfaCredentials: (accessToken: string) => Promise<CredentialsResponse>
   listWebAuthnDevices: (accessToken: string) => Promise<DeviceCredential[]>
   loginFromSession: (options?: AuthOptions) => Promise<void>
   loginWithCredentials: (params: LoginWithCredentialsParams) => Promise<AuthResult>
-  loginWithCustomToken: (params: { token: string; auth: AuthOptions }) => Promise<void>
+  loginWithCustomToken: (params: LoginWithCustomTokenParams) => Promise<void>
   loginWithPassword: (params: LoginWithPasswordParams) => Promise<AuthResult>
   loginWithSocialProvider: (provider: string, options?: AuthOptions) => Promise<void | InAppBrowser>
   loginWithWebAuthn: (params: LoginWithWebAuthnParams) => Promise<AuthResult>
-  logout: (params?: { redirectTo?: string; removeCredentials?: boolean }) => Promise<void>
+  logout: (params?: LogoutParams) => Promise<void>
   off: <K extends keyof Events>(eventName: K, listener: (payload: Events[K]) => void) => void
   on: <K extends keyof Events>(eventName: K, listener: (payload: Events[K]) => void) => void
   refreshTokens: (params: RefreshTokenParams) => Promise<AuthResult>
   remoteSettings: Promise<RemoteSettings>
+  removeMfaEmail: (params: RemoveMfaEmailParams) => Promise<void>
+  removeMfaPhoneNumber: (params: RemoveMfaPhoneNumberParams) => Promise<void>
   removeWebAuthnDevice: (accessToken: string, deviceId: string) => Promise<void>
   requestPasswordReset: (params: RequestPasswordResetParams) => Promise<void>
   sendEmailVerification: (params: EmailVerificationParams) => Promise<void>
@@ -70,25 +98,23 @@ export type Client = {
   signup: (params: SignupParams) => Promise<AuthResult>
   signupWithWebAuthn: (params: SignupWithWebAuthnParams, auth?: AuthOptions) => Promise<AuthResult>
   startMfaEmailRegistration: (params: StartMfaEmailRegistrationParams) => Promise<StartMfaEmailRegistrationResponse>
-  startMfaPhoneNumberRegistration: (params: StartMfaPhoneNumberRegistrationParams) => Promise<void>
+  startMfaPhoneNumberRegistration: (
+    params: StartMfaPhoneNumberRegistrationParams
+  ) => Promise<StartMfaPhoneNumberRegistrationResponse>
   startPasswordless: (
-    params: PasswordlessParams,
+    params: SingleFactorPasswordlessParams,
     options?: Omit<AuthOptions, 'useWebMessage'>
   ) => Promise<PasswordlessResponse>
-  unlink: (params: { accessToken: string; identityId: string; fields?: string }) => Promise<void>
+  unlink: (params: UnlinkParams) => Promise<void>
   updateEmail: (params: UpdateEmailParams) => Promise<void>
   updatePassword: (params: UpdatePasswordParams) => Promise<void>
-  updatePhoneNumber: (params: { accessToken: string; phoneNumber: string }) => Promise<void>
-  updateProfile: (params: { accessToken: string; redirectUrl?: string; data: Profile }) => Promise<void>
-  verifyPasswordless: (params: VerifyPasswordlessParams) => Promise<void>
-  verifyMfaPasswordless: (params: VerifyMfaPasswordlessParams) => Promise<AuthResult>
+  updatePhoneNumber: (params: UpdatePhoneNumberParams) => Promise<void>
+  updateProfile: (params: UpdateProfileParams) => Promise<void>
   verifyMfaEmailRegistration: (params: VerifyMfaEmailRegistrationParams) => Promise<void>
+  verifyMfaPasswordless: (params: VerifyMfaPasswordlessParams) => Promise<AuthResult>
   verifyMfaPhoneNumberRegistration: (params: VerifyMfaPhoneNumberRegistrationParams) => Promise<void>
-  verifyPhoneNumber: (params: { accessToken: string; phoneNumber: string; verificationCode: string }) => Promise<void>
-  getMfaStepUpToken: (params: StepUpParams) => Promise<StepUpResponse>
-  listMfaCredentials: (accessToken: string) => Promise<MfaCredentialsResponse>
-  removeMfaPhoneNumber: (params: RemoveMfaPhoneNumberParams) => Promise<void>
-  removeMfaEmail: (params: RemoveMfaEmailParams) => Promise<void>
+  verifyPasswordless: (params: VerifyPasswordlessParams) => Promise<void>
+  verifyPhoneNumber: (params: VerifyPhoneNumberParams) => Promise<void>
 }
 
 function checkParam<T>(data: T, key: keyof T) {
@@ -105,7 +131,14 @@ export function createClient(creationConfig: Config): Client {
   const { domain, clientId, language } = creationConfig
 
   const eventManager = createEventManager()
+
   const urlParser = createUrlParser(eventManager)
+
+  initCordovaCallbackIfNecessary(urlParser)
+
+  const baseUrl = `https://${domain}`
+
+  const baseIdentityUrl = `${baseUrl}/identity/v1`
 
   const remoteSettings = rawRequest<RemoteSettings>(
     `https://${domain}/identity/v1/config?${toQueryString({
@@ -114,24 +147,53 @@ export function createClient(creationConfig: Config): Client {
     })}`
   )
 
-  const apiClient = remoteSettings.then(
-    remoteConfig =>
-      new ApiClient({
-        config: {
-          ...creationConfig,
-          ...remoteConfig,
-        },
+  const apiClients = remoteSettings.then(remoteConfig => {
+    const { language, sso } = remoteConfig
+
+    const config = {
+      clientId,
+      baseUrl,
+      ...remoteConfig,
+    }
+
+    const http = createHttpClient({
+      baseUrl: baseIdentityUrl,
+      language,
+      acceptCookies: sso,
+    })
+
+    const oAuthClient = new OAuthClient({
+      config,
+      http,
+      eventManager,
+    })
+
+    return {
+      oAuth: oAuthClient,
+      mfa: new MfaClient({
+        http,
+        oAuthClient,
+      }),
+      webAuthn: new WebAuthnClient({
+        config,
+        http,
         eventManager,
-        urlParser,
-      })
-  )
+        oAuthClient,
+      }),
+      profile: new ProfileClient({
+        config,
+        http,
+        eventManager,
+      }),
+    }
+  })
 
   function addNewWebAuthnDevice(accessToken: string, friendlyName?: string) {
-    return apiClient.then(api => api.addNewWebAuthnDevice(accessToken, friendlyName))
+    return apiClients.then(clients => clients.webAuthn.addNewWebAuthnDevice(accessToken, friendlyName))
   }
 
   function checkSession(options: AuthOptions = {}) {
-    return apiClient.then(api => api.checkSession(options))
+    return apiClients.then(clients => clients.oAuth.checkSession(options))
   }
 
   function checkUrlFragment(url: string = window.location.href): boolean {
@@ -143,51 +205,59 @@ export function createClient(creationConfig: Config): Client {
   }
 
   function exchangeAuthorizationCodeWithPkce(params: TokenRequestParameters) {
-    return apiClient.then(api => api.exchangeAuthorizationCodeWithPkce(params))
+    return apiClients.then(clients => clients.oAuth.exchangeAuthorizationCodeWithPkce(params))
+  }
+
+  function getMfaStepUpToken(params: StepUpParams) {
+    return apiClients.then(clients => clients.mfa.getMfaStepUpToken(params))
   }
 
   function getSessionInfo() {
-    return apiClient.then(api => api.getSessionInfo())
+    return apiClients.then(clients => clients.oAuth.getSessionInfo())
   }
 
   function getSignupData(signupToken: string) {
-    return apiClient.then(api => api.getSignupData(signupToken))
+    return apiClients.then(clients => clients.profile.getSignupData(signupToken))
   }
 
-  function getUser(params: { accessToken: string; fields?: string }) {
-    return apiClient.then(api => api.getUser(params))
+  function getUser(params: GetUserParams) {
+    return apiClients.then(clients => clients.profile.getUser(params))
+  }
+
+  function listMfaCredentials(accessToken: string) {
+    return apiClients.then(clients => clients.mfa.listMfaCredentials(accessToken))
   }
 
   function listWebAuthnDevices(accessToken: string) {
-    return apiClient.then(api => api.listWebAuthnDevices(accessToken))
+    return apiClients.then(clients => clients.webAuthn.listWebAuthnDevices(accessToken))
   }
 
   function loginFromSession(options: AuthOptions = {}) {
-    return apiClient.then(api => api.loginFromSession(options))
+    return apiClients.then(clients => clients.oAuth.loginFromSession(options))
   }
 
   function loginWithCredentials(params: LoginWithCredentialsParams) {
-    return apiClient.then(api => api.loginWithCredentials(params))
+    return apiClients.then(clients => clients.oAuth.loginWithCredentials(params))
   }
 
-  function loginWithCustomToken(params: { token: string; auth: AuthOptions }) {
-    return apiClient.then(api => api.loginWithCustomToken(params))
+  function loginWithCustomToken(params: LoginWithCustomTokenParams) {
+    return apiClients.then(clients => clients.oAuth.loginWithCustomToken(params))
   }
 
   function loginWithPassword(params: LoginWithPasswordParams) {
-    return apiClient.then(api => api.loginWithPassword(params))
+    return apiClients.then(clients => clients.oAuth.loginWithPassword(params))
   }
 
   function loginWithSocialProvider(provider: string, options: AuthOptions = {}) {
-    return apiClient.then(api => api.loginWithSocialProvider(provider, options))
+    return apiClients.then(clients => clients.oAuth.loginWithSocialProvider(provider, options))
   }
 
   function loginWithWebAuthn(params: LoginWithWebAuthnParams) {
-    return apiClient.then(api => api.loginWithWebAuthn(params))
+    return apiClients.then(clients => clients.webAuthn.loginWithWebAuthn(params))
   }
 
-  function logout(params: { redirectTo?: string; removeCredentials?: boolean } = {}) {
-    return apiClient.then(api => api.logout(params))
+  function logout(params: LogoutParams = {}) {
+    return apiClients.then(clients => clients.oAuth.logout(params))
   }
 
   function off<K extends keyof Events>(eventName: K, listener: (payload: Events[K]) => void): void {
@@ -205,99 +275,91 @@ export function createClient(creationConfig: Config): Client {
   }
 
   function refreshTokens(params: RefreshTokenParams) {
-    return apiClient.then(api => api.refreshTokens(params))
-  }
-
-  function removeWebAuthnDevice(accessToken: string, deviceId: string) {
-    return apiClient.then(api => api.removeWebAuthnDevice(accessToken, deviceId))
-  }
-
-  function requestPasswordReset(params: RequestPasswordResetParams) {
-    return apiClient.then(api => api.requestPasswordReset(params))
-  }
-
-  function sendEmailVerification(params: EmailVerificationParams) {
-    return apiClient.then(api => api.sendEmailVerification(params))
-  }
-
-  function sendPhoneNumberVerification(params: PhoneNumberVerificationParams) {
-    return apiClient.then(api => api.sendPhoneNumberVerification(params))
-  }
-
-  function signup(params: SignupParams) {
-    return apiClient.then(api => api.signup(params))
-  }
-
-  function signupWithWebAuthn(params: SignupWithWebAuthnParams, auth?: AuthOptions) {
-    return apiClient.then(api => api.signupWithWebAuthn(params, auth))
-  }
-
-  function startMfaEmailRegistration(params: StartMfaEmailRegistrationParams) {
-    return apiClient.then(api => api.startMfaEmailRegistration(params))
-  }
-
-  function startMfaPhoneNumberRegistration(params: StartMfaPhoneNumberRegistrationParams) {
-    return apiClient.then(api => api.startMfaPhoneNumberRegistration(params))
-  }
-
-  function startPasswordless(params: PasswordlessParams, options: AuthOptions = {}) {
-    return apiClient.then(api => api.startPasswordless(params, options))
-  }
-
-  function unlink(params: { accessToken: string; identityId: string; fields?: string }) {
-    return apiClient.then(api => api.unlink(params))
-  }
-
-  function updateEmail(params: UpdateEmailParams) {
-    return apiClient.then(api => api.updateEmail(params))
-  }
-
-  function updatePassword(params: UpdatePasswordParams) {
-    return apiClient.then(api => api.updatePassword(params))
-  }
-
-  function updatePhoneNumber(params: { accessToken: string; phoneNumber: string }) {
-    return apiClient.then(api => api.updatePhoneNumber(params))
-  }
-
-  function updateProfile(params: { accessToken: string; redirectUrl?: string; data: Profile }) {
-    return apiClient.then(api => api.updateProfile(params))
-  }
-
-  function verifyMfaPasswordless(params: VerifyMfaPasswordlessParams) {
-    return apiClient.then(api => api.verifyMfaPasswordless(params))
-  }
-
-  function verifyMfaEmailRegistration(params: VerifyMfaEmailRegistrationParams) {
-    return apiClient.then(api => api.verifyMfaEmailRegistration(params))
-  }
-
-  function verifyMfaPhoneNumberRegistration(params: VerifyMfaPhoneNumberRegistrationParams) {
-    return apiClient.then(api => api.verifyMfaPhoneNumberRegistration(params))
-  }
-
-  function verifyPasswordless(params: VerifyPasswordlessParams, auth?: AuthOptions) {
-    return apiClient.then(api => api.verifyPasswordless(params, auth))
-  }
-
-  function verifyPhoneNumber(params: { accessToken: string; phoneNumber: string; verificationCode: string }) {
-    return apiClient.then(api => api.verifyPhoneNumber(params))
-  }
-
-  function getMfaStepUpToken(params: StepUpParams) {
-    return apiClient.then(api => api.getMfaStepUpToken(params))
-  }
-
-  function listMfaCredentials(accessToken: string) {
-    return apiClient.then(api => api.listMfaCredentials(accessToken))
-  }
-
-  function removeMfaPhoneNumber(params: RemoveMfaPhoneNumberParams) {
-    return apiClient.then(api => api.removeMfaPhoneNumber(params))
+    return apiClients.then(clients => clients.oAuth.refreshTokens(params))
   }
 
   function removeMfaEmail(params: RemoveMfaEmailParams) {
-    return apiClient.then(api => api.removeMfaEmail(params))
+    return apiClients.then(clients => clients.mfa.removeMfaEmail(params))
+  }
+
+  function removeMfaPhoneNumber(params: RemoveMfaPhoneNumberParams) {
+    return apiClients.then(clients => clients.mfa.removeMfaPhoneNumber(params))
+  }
+
+  function removeWebAuthnDevice(accessToken: string, deviceId: string) {
+    return apiClients.then(clients => clients.webAuthn.removeWebAuthnDevice(accessToken, deviceId))
+  }
+
+  function requestPasswordReset(params: RequestPasswordResetParams) {
+    return apiClients.then(clients => clients.profile.requestPasswordReset(params))
+  }
+
+  function sendEmailVerification(params: EmailVerificationParams) {
+    return apiClients.then(clients => clients.profile.sendEmailVerification(params))
+  }
+
+  function sendPhoneNumberVerification(params: PhoneNumberVerificationParams) {
+    return apiClients.then(clients => clients.profile.sendPhoneNumberVerification(params))
+  }
+
+  function signup(params: SignupParams) {
+    return apiClients.then(clients => clients.oAuth.signup(params))
+  }
+
+  function signupWithWebAuthn(params: SignupWithWebAuthnParams, auth?: AuthOptions) {
+    return apiClients.then(clients => clients.webAuthn.signupWithWebAuthn(params, auth))
+  }
+
+  function startMfaEmailRegistration(params: StartMfaEmailRegistrationParams) {
+    return apiClients.then(clients => clients.mfa.startMfaEmailRegistration(params))
+  }
+
+  function startMfaPhoneNumberRegistration(params: StartMfaPhoneNumberRegistrationParams) {
+    return apiClients.then(clients => clients.mfa.startMfaPhoneNumberRegistration(params))
+  }
+
+  function startPasswordless(params: SingleFactorPasswordlessParams, options: AuthOptions = {}) {
+    return apiClients.then(clients => clients.oAuth.startPasswordless(params, options))
+  }
+
+  function unlink(params: UnlinkParams) {
+    return apiClients.then(clients => clients.profile.unlink(params))
+  }
+
+  function updateEmail(params: UpdateEmailParams) {
+    return apiClients.then(clients => clients.profile.updateEmail(params))
+  }
+
+  function updatePassword(params: UpdatePasswordParams) {
+    return apiClients.then(clients => clients.profile.updatePassword(params))
+  }
+
+  function updatePhoneNumber(params: UpdatePhoneNumberParams) {
+    return apiClients.then(clients => clients.profile.updatePhoneNumber(params))
+  }
+
+  function updateProfile(params: UpdateProfileParams) {
+    return apiClients.then(clients => clients.profile.updateProfile(params))
+  }
+
+  function verifyMfaEmailRegistration(params: VerifyMfaEmailRegistrationParams) {
+    return apiClients.then(clients => clients.mfa.verifyMfaEmailRegistration(params))
+  }
+
+  function verifyMfaPasswordless(params: VerifyMfaPasswordlessParams) {
+    return apiClients.then(clients => clients.mfa.verifyMfaPasswordless(params))
+  }
+
+  function verifyMfaPhoneNumberRegistration(params: VerifyMfaPhoneNumberRegistrationParams) {
+    return apiClients.then(clients => clients.mfa.verifyMfaPhoneNumberRegistration(params))
+  }
+
+  function verifyPasswordless(params: VerifyPasswordlessParams, auth?: AuthOptions) {
+    return apiClients.then(clients => clients.oAuth.verifyPasswordless(params, auth))
+  }
+
+  function verifyPhoneNumber(params: VerifyPhoneNumberParams) {
+    return apiClients.then(clients => clients.profile.verifyPhoneNumber(params))
   }
 
   return {
@@ -305,9 +367,11 @@ export function createClient(creationConfig: Config): Client {
     checkSession,
     checkUrlFragment,
     exchangeAuthorizationCodeWithPkce,
+    getMfaStepUpToken,
     getSessionInfo,
     getSignupData,
     getUser,
+    listMfaCredentials,
     listWebAuthnDevices,
     loginFromSession,
     loginWithCredentials,
@@ -320,6 +384,8 @@ export function createClient(creationConfig: Config): Client {
     on,
     refreshTokens,
     remoteSettings,
+    removeMfaEmail,
+    removeMfaPhoneNumber,
     removeWebAuthnDevice,
     requestPasswordReset,
     sendEmailVerification,
@@ -334,14 +400,10 @@ export function createClient(creationConfig: Config): Client {
     updatePassword,
     updatePhoneNumber,
     updateProfile,
-    verifyPasswordless,
-    verifyMfaPasswordless,
     verifyMfaEmailRegistration,
+    verifyMfaPasswordless,
     verifyMfaPhoneNumberRegistration,
+    verifyPasswordless,
     verifyPhoneNumber,
-    getMfaStepUpToken,
-    listMfaCredentials,
-    removeMfaPhoneNumber,
-    removeMfaEmail,
   }
 }