+++ src/main/webAuthnClient.ts
@@ -1,0 +1,184 @@
+import { AuthOptions } from './authOptions'
+import { AuthResult } from './authResult'
+import { HttpClient } from './httpClient'
+import { IdentityEventManager } from './identityEventManager'
+import { ApiClientConfig } from './main'
+import { AuthenticationToken } from './models'
+import OAuthClient from './oAuthClient'
+import { resolveScope } from './scopeHelper'
+import {
+  CredentialRequestOptionsSerialized,
+  DeviceCredential,
+  EmailLoginWithWebAuthnParams,
+  encodePublicKeyCredentialCreationOptions,
+  encodePublicKeyCredentialRequestOptions,
+  LoginWithWebAuthnParams,
+  PhoneNumberLoginWithWebAuthnParams,
+  publicKeyCredentialType,
+  RegistrationOptions,
+  serializeAuthenticationPublicKeyCredential,
+  serializeRegistrationPublicKeyCredential,
+  SignupWithWebAuthnParams,
+} from './webAuthnService'
+
+/**
+ * Identity Rest API Client
+ */
+export default class WebAuthnClient {
+  private config: ApiClientConfig
+  private http: HttpClient
+  private eventManager: IdentityEventManager
+  private oAuthClient: OAuthClient
+
+  private authenticationOptionsUrl = '/webauthn/authentication-options'
+  private authenticationUrl = '/webauthn/authentication'
+  private registrationOptionsUrl = '/webauthn/registration-options'
+  private readonly registrationUrl = '/webauthn/registration'
+  private readonly signupOptionsUrl = '/webauthn/signup-options'
+  private signupUrl = '/webauthn/signup'
+
+  constructor(props: {
+    config: ApiClientConfig
+    http: HttpClient
+    eventManager: IdentityEventManager
+    oAuthClient: OAuthClient
+  }) {
+    this.config = props.config
+    this.http = props.http
+    this.eventManager = props.eventManager
+    this.oAuthClient = props.oAuthClient
+
+    this.authenticationOptionsUrl = '/webauthn/authentication-options'
+    this.authenticationUrl = '/webauthn/authentication'
+    this.registrationOptionsUrl = '/webauthn/registration-options'
+    this.registrationUrl = '/webauthn/registration'
+    this.signupOptionsUrl = '/webauthn/signup-options'
+    this.signupUrl = '/webauthn/signup'
+  }
+
+  addNewWebAuthnDevice(accessToken: string, friendlyName?: string): Promise<void> {
+    if (window.PublicKeyCredential) {
+      const body = {
+        origin: window.location.origin,
+        friendlyName: friendlyName || window.navigator.platform,
+      }
+
+      return this.http
+        .post<RegistrationOptions>(this.registrationOptionsUrl, { body, accessToken })
+        .then(response => {
+          const publicKey = encodePublicKeyCredentialCreationOptions(response.options.publicKey)
+
+          return navigator.credentials.create({ publicKey })
+        })
+        .then(credentials => {
+          if (!credentials || credentials.type !== publicKeyCredentialType) {
+            return Promise.reject(new Error('Unable to register invalid public key credentials.'))
+          }
+
+          const serializedCredentials = serializeRegistrationPublicKeyCredential(credentials)
+
+          return this.http.post<void>(this.registrationUrl, { body: { ...serializedCredentials }, accessToken })
+        })
+        .catch(err => {
+          if (err.error) this.eventManager.fireEvent('login_failed', err)
+
+          return Promise.reject(err)
+        })
+    } else {
+      return Promise.reject(new Error('Unsupported WebAuthn API'))
+    }
+  }
+
+  listWebAuthnDevices(accessToken: string): Promise<DeviceCredential[]> {
+    return this.http.get<DeviceCredential[]>(this.registrationUrl, { accessToken })
+  }
+
+  loginWithWebAuthn(params: LoginWithWebAuthnParams): Promise<AuthResult> {
+    if (window.PublicKeyCredential) {
+      const body = {
+        clientId: this.config.clientId,
+        origin: window.location.origin,
+        scope: resolveScope(params.auth, this.config.scope),
+        email: (params as EmailLoginWithWebAuthnParams).email,
+        phoneNumber: (params as PhoneNumberLoginWithWebAuthnParams).phoneNumber,
+      }
+
+      return this.http
+        .post<CredentialRequestOptionsSerialized>(this.authenticationOptionsUrl, { body })
+        .then(response => {
+          const options = encodePublicKeyCredentialRequestOptions(response.publicKey)
+
+          return navigator.credentials.get({ publicKey: options })
+        })
+        .then(credentials => {
+          if (!credentials || credentials.type !== publicKeyCredentialType) {
+            return Promise.reject(new Error('Unable to authenticate with invalid public key credentials.'))
+          }
+
+          const serializedCredentials = serializeAuthenticationPublicKeyCredential(credentials)
+
+          return this.http
+            .post<AuthenticationToken>(this.authenticationUrl, { body: { ...serializedCredentials } })
+            .then(tkn => this.oAuthClient.loginCallback(tkn, params.auth))
+        })
+        .catch(err => {
+          if (err.error) this.eventManager.fireEvent('login_failed', err)
+
+          return Promise.reject(err)
+        })
+    } else {
+      return Promise.reject(new Error('Unsupported WebAuthn API'))
+    }
+  }
+
+  removeWebAuthnDevice(accessToken: string, deviceId: string): Promise<void> {
+    return this.http.remove<void>(`${this.registrationUrl}/${deviceId}`, { accessToken })
+  }
+
+  signupWithWebAuthn(params: SignupWithWebAuthnParams, auth?: AuthOptions): Promise<AuthResult> {
+    if (window.PublicKeyCredential) {
+      const body = {
+        origin: window.location.origin,
+        clientId: this.config.clientId,
+        friendlyName: params.friendlyName || window.navigator.platform,
+        profile: params.profile,
+        scope: resolveScope(auth, this.config.scope),
+        redirectUrl: params.redirectUrl,
+        returnToAfterEmailConfirmation: params.returnToAfterEmailConfirmation,
+      }
+
+      const registrationOptionsPromise = this.http.post<RegistrationOptions>(this.signupOptionsUrl, { body })
+
+      const credentialsPromise = registrationOptionsPromise.then(response => {
+        const publicKey = encodePublicKeyCredentialCreationOptions(response.options.publicKey)
+
+        return navigator.credentials.create({ publicKey })
+      })
+
+      return Promise.all([registrationOptionsPromise, credentialsPromise])
+        .then(([registrationOptions, credentials]) => {
+          if (!credentials || credentials.type !== publicKeyCredentialType) {
+            return Promise.reject(new Error('Unable to register invalid public key credentials.'))
+          }
+
+          const serializedCredentials = serializeRegistrationPublicKeyCredential(credentials)
+
+          return this.http
+            .post<AuthenticationToken>(this.signupUrl, {
+              body: {
+                publicKeyCredential: serializedCredentials,
+                webauthnId: registrationOptions.options.publicKey.user.id,
+              },
+            })
+            .then(tkn => this.oAuthClient.loginCallback(tkn, auth))
+        })
+        .catch(err => {
+          if (err.error) this.eventManager.fireEvent('login_failed', err)
+
+          return Promise.reject(err)
+        })
+    } else {
+      return Promise.reject(new Error('Unsupported WebAuthn API'))
+    }
+  }
+}