+++ CHANGELOG.md
@@ -8,7 +8,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Changed
-Replaced tslint by eslint.
+
+- Align `startMfaPhoneNumberRegistration` response on the `startMfaEmailRegistration` one.
+- Replaced tslint by eslint.
 
 ## [1.22.0] - 2022-02-22
 
@@ -24,7 +26,8 @@ Replaced tslint by eslint.
 and [requestPasswordReset](https://developer.reachfive.com/sdk-core/requestPasswordReset.html)
 
 ### Changed
-- Support the step up from a fresh access token with function `getMfaStepUpToken`
+
+Support the step up from a fresh access token with function `getMfaStepUpToken`
 
 ## [1.21.1] - 2021-11-30